sell-master/src/main/java/com/cheng/sell/service/impl/SellerServiceImpl.java
package com.cheng.sell.service.impl;

import com.cheng.sell.dataobject.SellerInfo;
import com.cheng.sell.repository.SellerInfoRepository;
import com.cheng.sell.service.SellerService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

/**
 * Created with IntelliJ IDEA.
 * Description:
 *
 * @author cheng
 * Date: 2018-07-12
 * Time: 上午11:18
 */
@Service
public class SellerServiceImpl implements SellerService {

    @Autowired
    private SellerInfoRepository repository;

    @Override
    public SellerInfo findSellerInfoByOpenid(String openid) {
        return repository.findByOpenid(openid);
    }

}
